Add config option to hide project on preserve
The preserve option as it is today is only of limited use. It deletes
the project from Gerrit, but preserve the Git repository. As result
the project will reappear once the Gerrit server is restarted.

Add a configuration option to instead hide the project when the
preserve option is used. Hiding the project means that the project
state is set to 'HIDDEN', all access rights are removed and the
project is reparented to the 'Deleted-Projects' project. The
'Deleted-Projects' project is automatically created under the root
project when it does not exist yet. This way the project disappears
for normal users and administrators can still recover it. The
reparenting is needed to remove any inherited access rights and it's
also nice for administrators to find all deleted projects under a
common parent.

The default behavior is not changed.

+++ src/main/java/com/googlesource/gerrit/plugins/deleteproject/DeleteProject.java
@@ -29,6 +29,7 @@ import com.google.gerrit.server.IdentifiedUser;
 import com.google.gerrit.server.account.CapabilityControl;
 import com.google.gerrit.server.config.AllProjectsName;
 import com.google.gerrit.server.config.AllProjectsNameProvider;
+import com.google.gerrit.server.config.PluginConfigFactory;
 import com.google.gerrit.server.project.ProjectResource;
 import com.google.gwtorm.server.OrmException;
 import com.google.inject.Inject;
@@ -59,6 +60,8 @@ class DeleteProject implements RestModifyView<ProjectResource, Input> {
   private final Provider<CurrentUser> userProvider;
   private final String pluginName;
   private final DeleteLog deleteLog;
+  private final PluginConfigFactory cfgFactory;
+  private final HideProject hideProject;
 
   @Inject
   DeleteProject(AllProjectsNameProvider allProjectsNameProvider,
@@ -68,7 +71,9 @@ class DeleteProject implements RestModifyView<ProjectResource, Input> {
       ProjectConfigDeleteHandler pcHandler,
       Provider<CurrentUser> userProvider,
       @PluginName String pluginName,
-      DeleteLog deleteLog) {
+      DeleteLog deleteLog,
+      PluginConfigFactory cfgFactory,
+      HideProject hideProject) {
     this.allProjectsName = allProjectsNameProvider.get();
     this.dbHandler = dbHandler;
     this.fsHandler = fsHandler;
@@ -77,6 +82,8 @@ class DeleteProject implements RestModifyView<ProjectResource, Input> {
     this.userProvider = userProvider;
     this.pluginName = pluginName;
     this.deleteLog = deleteLog;
+    this.cfgFactory = cfgFactory;
+    this.hideProject = hideProject;
   }
 
   @Override
@@ -129,18 +136,25 @@ class DeleteProject implements RestModifyView<ProjectResource, Input> {
     return dbHandler.getWarnings(rsrc.getControl().getProject());
   }
 
-  public void doDelete(ProjectResource rsrc, Input input)
-      throws OrmException, IOException, ResourceNotFoundException {
+  public void doDelete(ProjectResource rsrc, Input input) throws OrmException,
+      IOException, ResourceNotFoundException, ResourceConflictException {
     Project project = rsrc.getControl().getProject();
+    boolean preserve = input != null && input.preserve;
     Exception ex = null;
     try {
-      dbHandler.delete(project);
-      try {
-        fsHandler.delete(project, input == null ? false : input.preserve);
-      } catch (RepositoryNotFoundException e) {
-        throw new ResourceNotFoundException();
+      if (!preserve
+          || !cfgFactory.getFromGerritConfig(pluginName)
+              .getBoolean("hideProjectOnPreserve", false)) {
+        dbHandler.delete(project);
+        try {
+          fsHandler.delete(project, preserve);
+        } catch (RepositoryNotFoundException e) {
+          throw new ResourceNotFoundException();
+        }
+        cacheHandler.delete(project);
+      } else {
+        hideProject.apply(rsrc);
       }
-      cacheHandler.delete(project);
     } catch (Exception e) {
       ex = e;
       throw e;
